chore(team): split mailbox store mutations
