+++ src/team/manager.rs
@@ -18,6 +18,7 @@ mod mailbox_shared_thread;
 mod mailbox_sqlite;
 mod mailbox_store;
 mod mailbox_store_inbox;
+mod mailbox_store_mutations;
 mod mailbox_store_relay;
 #[cfg(test)]
 mod mailbox_tests;
